requirements.txt
torch
torchvision
numpy
scikit-image
matplotlib
tqdm
Pillow